优化浮点数比较，使用 fabsl 替代 fabs；调整 NoneType 为类并修复格式；更新 str 类的输出操作符声明；添加 typeinfo 的默认构造函数和析构函数。

+++ floating_point.cpp
@@ -1,5 +1,6 @@
 #include "floating_point.h"
 
+#include <cmath>
 #include <iostream>
 
 #include "dict.h"
@@ -21,7 +22,7 @@ std::ostream& operator<<(std::ostream& os, const float_& f) {
 }
 
 bool operator==(const float_& lhs, const float_& rhs) {
-  return std::fabs(lhs.value() - rhs.value()) < float_::epsilon().value();
+  return std::fabsl(lhs.value() - rhs.value()) < float_::epsilon().value();
 }
 
 bool operator<(const float_& lhs, const float_& rhs) {
